[feat] move logger to Common
