+++ common/src/main/java/com/samsung/file/FileConstants.java
@@ -1,4 +1,4 @@
-package com.samsung;
+package com.samsung.file;
 
 public final class FileConstants {
     private FileConstants() {}
