+++ common/src/main/java/com/samsung/log/LogCompressingAppender.java
