+++ common/src/main/java/com/samsung/log/LogCompressor.java
