+++ common/src/test/java/com/samsung/file/FileManagerTest.java
@@ -1,4 +1,4 @@
-package com.samsung;
+package com.samsung.file;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
