+++ shell/src/main/java/com/samsung/command/testscript/ScriptManager.java
@@ -1,6 +1,6 @@
 package com.samsung.command.testscript;
 
-import com.samsung.FileManager;
+import com.samsung.file.FileManager;
 import com.samsung.file.JarExecutor;
 import lombok.extern.slf4j.Slf4j;
