+++ shell/src/main/java/com/samsung/command/testshell/TestShellManager.java
@@ -1,6 +1,6 @@
 package com.samsung.command.testshell;
 
-import com.samsung.FileManager;
+import com.samsung.file.FileManager;
 import com.samsung.file.JarExecutor;
 import lombok.extern.slf4j.Slf4j;
 import java.util.List;
