+++ ssd/src/test/java/com/samsung/SSDManagerTest.java
@@ -1,5 +1,6 @@
 package com.samsung;
 
+import com.samsung.file.FileManager;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
